chores(telegram): repackaged tests
