+++ modules/telegram/src/test/java/io/starter/units/handler/BaseTest.java
@@ -1,0 +1,8 @@
+package io.starter.units.handler;
+
+import net.datafaker.Faker;
+
+public abstract class BaseTest {
+
+  protected final Faker faker = new Faker();
+}
