+++ modules/telegram/src/test/java/io/starter/units/handler/callback/skills/InteractionsTest.java
@@ -1,31 +1,20 @@
-package io.starter.units.handler;
+package io.starter.units.handler.callback.skills;
 
 import java.util.List;
 import java.util.stream.IntStream;
 
 import io.starter.dataproviders.CallbackHandlerProvider;
-import io.starter.telegram.cash.CallbackCash;
-import io.starter.telegram.cash.MessageCash;
 import io.starter.telegram.cash.state.CallbackState;
 import io.starter.telegram.constants.Constants;
-import io.starter.telegram.dao.SkillDao;
-import io.starter.telegram.dao.UserDao;
 import io.starter.telegram.entity.LeagueEntity;
 import io.starter.telegram.handler.UpdateHandler;
 import io.starter.telegram.model.aggregator.Skill;
 import io.starter.telegram.model.telegram.TelegramFacade;
-import io.starter.telegram.service.CallbackAnswerService;
-import io.starter.telegram.service.MessageAnswerService;
+import io.starter.units.handler.callback.BaseCallbackTest;
 
-import net.datafaker.Faker;
 import org.telegram.telegrambots.meta.api.methods.AnswerCallbackQuery;
 import org.telegram.telegrambots.meta.api.methods.botapimethods.BotApiMethod;
 import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageText;
-import org.telegram.telegrambots.meta.api.objects.CallbackQuery;
-import org.telegram.telegrambots.meta.api.objects.Update;
-import org.telegram.telegrambots.meta.api.objects.User;
-import org.telegram.telegrambots.meta.api.objects.message.MaybeInaccessibleMessage;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -34,31 +23,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 
-public class CallbackHandlerTest {
-
-  private final SkillDao skillDao = mock(SkillDao.class);
-  private final UserDao userDao = mock(UserDao.class);
-  private final MaybeInaccessibleMessage message = mock(MaybeInaccessibleMessage.class);
-  private final CallbackQuery callbackQuery = mock(CallbackQuery.class);
-  private final Update update = mock(Update.class);
-  private final User user = mock(User.class);
-
-  private final CallbackAnswerService callbackAnswerService = spy(new CallbackAnswerService(skillDao, userDao));
-  private final MessageAnswerService messageAnswerService = spy(MessageAnswerService.class);
-  private final CallbackCash callbackCash = spy(CallbackCash.class);
-  private final MessageCash messageCash = spy(MessageCash.class);
-
-  private final Faker faker = new Faker();
-
-  @BeforeClass
-  void setUp() {
-    when(update.hasCallbackQuery()).thenReturn(true);
-    when(update.getCallbackQuery()).thenReturn(callbackQuery);
-    when(callbackQuery.getFrom()).thenReturn(user);
-    when(callbackQuery.getMessage()).thenReturn(message);
-    doNothing().when(userDao).saveWhenNotExist(user);
-    doNothing().when(userDao).saveLastMessageId(user, message);
-  }
+public class InteractionsTest extends BaseCallbackTest {
 
   @Test(description = "Bot should react on clicking button 'skills'")
   void testWhenUserClickBtnSkills() {
