+++ modules/telegram/src/test/java/io/starter/units/handler/message/commands/CommandsTest.java
@@ -1,54 +1,22 @@
-package io.starter.units.handler;
+package io.starter.units.handler.message.commands;
 
-import io.starter.telegram.cash.CallbackCash;
-import io.starter.telegram.cash.MessageCash;
 import io.starter.telegram.constants.Constants;
-import io.starter.telegram.dao.UserDao;
 import io.starter.telegram.handler.UpdateHandler;
 import io.starter.telegram.model.telegram.TelegramFacade;
-import io.starter.telegram.service.CallbackAnswerService;
-import io.starter.telegram.service.MessageAnswerService;
+import io.starter.units.handler.message.BaseMessageTest;
 
-import net.datafaker.Faker;
 import org.telegram.telegrambots.meta.api.methods.botapimethods.BotApiMethod;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
-import org.telegram.telegrambots.meta.api.objects.Update;
-import org.telegram.telegrambots.meta.api.objects.User;
-import org.telegram.telegrambots.meta.api.objects.message.Message;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 import static io.starter.telegram.cash.state.MessageState.SETTINGS;
 import static io.starter.telegram.cash.state.MessageState.START;
 import static io.starter.telegram.cash.state.MessageState.WELCOME;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.doNothing;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 
-public class MessageHandlerTest {
-
-  private final CallbackAnswerService callbackAnswerService = mock(CallbackAnswerService.class);
-  private final UserDao userDao = mock(UserDao.class);
-  private final Message message = mock(Message.class);
-  private final Update update = mock(Update.class);
-  private final User user = mock(User.class);
-
-  private final MessageAnswerService messageAnswerService = spy(MessageAnswerService.class);
-  private final CallbackCash callbackCash = spy(CallbackCash.class);
-  private final MessageCash messageCash = spy(MessageCash.class);
-
-  private final Faker faker = new Faker();
-
-  @BeforeClass
-  void setUp() {
-    when(update.hasCallbackQuery()).thenReturn(false);
-    when(update.getMessage()).thenReturn(message);
-    when(message.getFrom()).thenReturn(user);
-    doNothing().when(userDao).saveWhenNotExist(user);
-    doNothing().when(userDao).saveLastMessageId(user, message);
-  }
+public class CommandsTest extends BaseMessageTest {
 
   @Test(description = "Bot should welcome user when he writes command '/start'")
   void testWhenBotGotCommandStart() {
